Refactor setting declarations into their own source file
